Updated css
Added dog & cat css for links
Added slight coloring to detail-content background and padding to
emphasize the content more.

Bugs:
The arrow starts close to the position instead of going from old row
position.  It does however nudge a tiny bit which is a bug but barely
noticeable unless you stare at it!